Route events through the room the character is in
Provides a way to override where the events will go, but by default will
get sent to the room the player is in. Adds a callback for the game's
code to actually process the events. Foreman can also route internal
events if something needs to delay an event (or have a timer.)