Push a lot of movement_request functionality higher up
Have the movement_request functions return a tuple of what it wants to
happen and let the room process that into the proper events. Also let
the room find the exit being requested to move to and pass it to the
callback function, since 99% of the time we want that.