Use the Meta protocol for vitals and refactor event text store
Let vitals slip through into exported events from the game

Meta.trim is a module that attaches the `__struct__` key with
`Meta.Trimmed` to tag it's been trimmed. When encoded into JSON, remove
the custom key.

Also refactors the `eventTextHandlers` to have a common function. This
keeps the keys known in the front end.